services/django-api/requirements.txt
# ============================================
# ForgeLink Django API - Dependencies
# ============================================

# Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0

# Admin UI
django-unfold==0.30.0

# GraphQL
graphene-django==3.2.2
django-graphql-jwt==0.4.0

# Database
psycopg[binary]==3.2.1

# TDengine
taospy==2.7.12

# Async tasks
celery[redis]==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Kafka
confluent-kafka==2.4.0

# RabbitMQ (via Celery)
# Already included with celery[redis]

# API clients
httpx==0.27.0
grpcio==1.64.1
grpcio-tools==1.64.1

# Authentication
PyJWT==2.8.0
cryptography==42.0.8

# Validation
pydantic==2.7.4
pydantic-settings==2.3.3

# Observability
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-celery==0.46b0
opentelemetry-instrumentation-redis==0.46b0
opentelemetry-instrumentation-psycopg==0.46b0
prometheus-client==0.20.0
django-prometheus==2.3.1
structlog==24.2.0

# AI/ML
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# Slack
slack-sdk==3.27.1

# Firebase (FCM)
firebase-admin==6.5.0

# Rate limiting
django-ratelimit==4.1.0

# Real-time WebSocket (Socket.IO)
python-socketio==5.11.2

# Utilities
python-dateutil==2.9.0
pytz==2024.1
python-dotenv==1.0.1

# Production server
gunicorn==22.0.0
uvicorn[standard]==0.30.1

# Development
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
factory-boy==3.3.0
faker==25.9.1
black==24.4.2
flake8==7.0.0
isort==5.13.2
mypy==1.10.0
django-stubs==5.0.2
